requirements.txt
Flask
Flask-SocketIO
Flask-SQLAlchemy
Flask-Caching
psycopg2-binary
python-dotenv
gunicorn
eventlet
Flask-Migrate
